Add ItemStorage
Here I'm introducing the concept of storage.
I don't want to have to create a database to store these for now unless 
I had to.

So what I'm doing here is representing how the items are in a storage 
facility before the customer buys them and give it a function to find 
those pre-stored items to be found by their item code.