Add Checkout with pricing rules
This commit implments the princing rules necessary for this code challenge.

I have discovered that the test data provided in the README.md file is
incorrect for the third case, hence living the comment on the last test in the checkout_spec.

I have also moved the rounding to two decimal numbers to the last total,
this prevents rounding up too many times (each rule) that the sum (or deduction) would become incorrect.