Update Rules::Discount and Rules::BulkDiscount
I have made corrections, due to a interpretation mistake.

The code was applying the discount to the summed price of all product 
and then removed the bulk discount. 

This is wrong, because in a real life scenario a product discount is 
applied immediately to the product and then any percentage discounts are 
applied at the end.

There was a significant difference of 15p.

This commit fixes that mistake and maintains the signature of each Rule 
class.