Add Rules::BuyOneGetOneFree
I decided to have some fun and add a new rule into the code challenge.

This rule allows people to buy an item and get a second item for free. 
It considers multiples of the minimum quantity to apply that discount.

Example: The minimum quantity for coffee is 2 (where one is free), but 
when a person buys 4 (a multiple of 2, then two are free). When it's 
only 3, only one coffee will be free.