Add Checkout test case for buy one get one free
This commit also introduces logic to move the Rules::Discount to be the 
last Rule to be interated, this removes the worry from the user to 
remember the order of which rule.

In the future another rule should be implememented to prevent an 
instance of a Checkout to receive more than one Rules::Discount.

If that does happen, an error should be raised as it is not possible to 
decide which discount should be in place.